Consolidate pypi_name_to_conda_name in lookup module
